# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Subscription Aggregation API
  version: 1.0.0
  description: REST API для CRUDL операций над онлайн-подписками и подсчета суммарной стоимости за период.
servers:
  - url: http://localhost:8080
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200': { description: OK }
  /api/v1/subscriptions:
    post:
      summary: Создать подписку
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateSubscriptionRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
    get:
      summary: Получить список подписок
      parameters:
        - in: query
          name: limit
          schema: { type: integer, default: 20, maximum: 100 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
        - in: query
          name: user_id
          schema: { type: string, format: uuid }
        - in: query
          name: service_name
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Subscription' }
  /api/v1/subscriptions/sum:
    get:
      summary: Подсчитать сумму подписок за период
      description: Сумма считается помесячно за месяцы пересечения подписки с периодом from-to включительно.
      parameters:
        - in: query
          name: from
          required: true
          schema: { type: string, example: "07-2025" }
        - in: query
          name: to
          required: true
          schema: { type: string, example: "09-2025" }
        - in: query
          name: user_id
          schema: { type: string, format: uuid }
        - in: query
          name: service_name
          schema: { type: string, example: "Yandex Plus" }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SumResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /api/v1/subscriptions/{id}:
    get:
      summary: Получить подписку по ID
      parameters:
        - $ref: '#/components/parameters/SubscriptionID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      summary: Обновить подписку
      parameters:
        - $ref: '#/components/parameters/SubscriptionID'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateSubscriptionRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Subscription' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      summary: Удалить подписку
      parameters:
        - $ref: '#/components/parameters/SubscriptionID'
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  parameters:
    SubscriptionID:
      in: path
      name: id
      required: true
      schema: { type: integer, format: int64 }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    CreateSubscriptionRequest:
      type: object
      required: [service_name, price, user_id, start_date]
      properties:
        service_name: { type: string, example: "Yandex Plus" }
        price: { type: integer, example: 400, minimum: 1 }
        user_id: { type: string, format: uuid, example: "60601fee-2bf1-4721-ae6f-7636e79a0cba" }
        start_date: { type: string, example: "07-2025" }
        end_date: { type: string, nullable: true, example: "12-2025" }
    UpdateSubscriptionRequest:
      type: object
      properties:
        service_name: { type: string, example: "Yandex Plus" }
        price: { type: integer, example: 400, minimum: 1 }
        user_id: { type: string, format: uuid }
        start_date: { type: string, example: "07-2025" }
        end_date: { type: string, nullable: true, example: "12-2025" }
    Subscription:
      allOf:
        - $ref: '#/components/schemas/CreateSubscriptionRequest'
        - type: object
          properties:
            id: { type: integer, format: int64, example: 1 }
            created_at: { type: string, format: date-time }
            updated_at: { type: string, format: date-time }
    SumResponse:
      type: object
      properties:
        total: { type: integer, example: 1200 }
        from: { type: string, example: "07-2025" }
        to: { type: string, example: "09-2025" }
        user_id: { type: string, format: uuid }
        service_name: { type: string }
    Error:
      type: object
      properties:
        error: { type: string, example: "invalid request" }